aiServices/requirements.txt
# HTTP & scraping
httpx[http2]==0.27.0
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2

# IA
openai==1.30.1

# Servidor API
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1

# Config
python-dotenv==1.0.1

# Tests
pytest==8.2.0
pytest-asyncio==0.23.7
